code/06-semantic-cache/requirements.txt
anthropic>=0.40.0
sentence-transformers>=2.7.0
redis>=5.0.0
python-dotenv>=1.0.0
